+++ UnityProject/Assets/GameScripts/HotFix/GameLogic/DataCenter/ClientSaveData/SystemSaveData.cs
@@ -1,10 +1,13 @@
+using Nino.Core;
+
 namespace GameLogic
 {
     /// <summary>
     /// 系统设置存档数据。
     /// </summary>
+    [NinoType(containNonPublicMembers: true)]
     [ClientSaveData("SystemSaveData")]
-    public sealed class SystemSaveData : BaseClientSaveData
+    public sealed partial class SystemSaveData : BaseClientSaveData
     {
         /// <summary>
         /// 系统保存数据类型枚举
@@ -17,7 +20,7 @@ namespace GameLogic
         /// <summary>
         /// 系统设置参数数组
         /// </summary>
-        public int[] SettingParams { get; private set; } = new int[(int)SaveType.Max];
+        public int[] SettingParams { get; internal set; } = new int[(int)SaveType.Max];
 
         /// <summary>
         /// 获取系统保存数据实例